fix: attempt to fix ZIP artifact download failure when using checkout@v3
The command that we used to resolve the current branch name started
returning "heads/branch" instead of the intended "branch" when using the
new major version of the GitHub Actions checkout action. This broke the
download of the previously generated ZIP file, negatively impacting
performance.

Let's use a newer Git command to get the branch name that might work
better. Note that we can't use the GITHUB_REF_NAME environment variable
because it'd be set to a tag name if the workflow is triggered on a tag.

While at it, introduce some error messages to better debug what's going
on.