tweak(ci): replace checkout@v3 test workflow by a double-checkout test
By pure luck, I've stumbled upon the osfanmuffin/muffinhunt-datapack
repository, which uses PackSquash to optimize a datapack, while looking
at the dependant graph. This project has an interesting PackSquash
optimization workflow that clones the repository twice, and doing so
breaks the action logic for downloading the latest ZIP file. I initially
thought the error was caused due to changes in the new checkout action
version, not noticing such a glaring weird thing.

Even though it makes no sense to clone a GitHub repository twice in a
workflow, the action should be as robust as possible and handle that
edge case as good as possible. Let's add tests for it, to see if
changing the command used to get the branch name helped.