fix(entrypoint): override PackSquash color and emoji logic
The GitHub action input parameters are already set to sane values,
considering this environment support for colored messages and emojis.
For those cases that the input parameters are set to false, it makes
sense to override any PackSquash default by setting the related
environment variables to empty values, instead of not setting them.