fix(entrypoint): rework artifact caching behavior
7c7b41491beb72df78323194d1bcdfbd7109a979 made me think that there might
be more things going wrong with caching, and that was true. The action
should attempt to use any cached system ID, as long as it was generated
for the same PackSquash options. And it makes little sense to download
the previous artifact if we have no previous or fixed system ID to read
it with.

Moreover, the branch filtering logic did not account properly for PRs,
falling back to no filtering for some PR events after
51e4f3c743e9c9da7044c6444235f27fe253a7a5. The best way to account for
PRs sanely I can come up with is to get the head branch for the current
run from the GitHub API and filter by that. Head branches are the target
branches for push events, but for PRs they are whatever name the PR
author decided in their fork or different repository branch, which is
fine for our purposes.

Some comments were added to better explain the (complicated) thought
process behind all of these changes.