tweak: overhaul options file generation. Tweak out path in custom ones
Overriding the output directory path present in custom options files is
necessary to effectively hide the implementation detail of the output
path in all cases. However, doing such a modification properly requires
parsing TOML files.

This is now very easy to do, as migrating to TypeScript allows us to
just pull in a TOML parser, instead of having to rely on external shell
commands or dumb string concatenations and splits that do not handle
edge cases properly.

Given that we need a proper TOML serializer and deserializer anyway,
let's use it to generate the options file too, which has the nice side
effect of being able to validate its correctness and data types within
TypeScript.

The documentation was updated accordingly, and a test case for custom
options files was added.