fix(submodules): only fail shallow check if the submodule is shallow
When the repository has submodules, it can happen that the main
repository is shallow, but its submodules have the full commit history.
Therefore, failing the shallow repository check when we only care about
the pack files in a submodule is overly restrictive.

Fix that by changing the shallow repository check to consider the pack
directory and not the workspace. Git is smart enough to find its
containing repository or submodule automatically.