fix(git_set_file_times): do not fail when there are no files. Refactors
When git log is called on changeTime, it may return empty output if
there are no files to get the time of. Moreover, due to a GitHub Actions
toolkit bug, output.stdout is undefined in this case.

To avoid special-casing that in changeTime and avoid that bug, let's
ignore repositories with no files in ls in the first place, which is
also more efficient.