docs: revamp README structure. Improve readability of parameter list
Experience has shown that some people like the illusion of saving time
by copying and pasting around stuff around without bothering to read
more in-depth documentation first, for a variety of good (and bad)
reasons. To please those impatient people, let's move the usage examples
to a more prominent place.

In addition, GitHub renders tables in a way that reading the input
parameters description is uncomfortable. Let's flatten that table to a
structure of paragraphs and titles for better readability.