fix(cache): take into account custom options for cacheMayBeUsed flag
Previously, when a custom options file was used, the internal
cacheMayBeUsed flag would be set. When set, this flag checks that the
repository clone is non-shallow, sets file modification times according
to commit times, and restores any previous artifact, in preparation for
it to be reused by PackSquash.

However, all that work goes to waste if the options file tells
PackSquash to not use previous results anyway. Let's be smarter and set
that flag after parsing the options, taking into account the values that
were set in that file.