chore: cleanup PackSquash < 0.3.1 support code remnants
I'm not interested in supporting older PackSquash versions on newer
versions of the action because it requires effort, complicates
maintenance, and it might backfire as an excuse for users to don't
upgrade, even if they lack good reasons to stay outdated. Unless those
older versions happen to be compatible and there are no reasons to drop
that support, of course.

Let's remove code whose only purpose was to deal with older PackSquash
versions that the action rejects anyway.