fix(problem_matcher): work around loop pattern quirk
For some reason the loop pattern must be the one that sets the problem
message, and this can only be known in advance by reading the runner
source code. Yikes!

This workaround has the unintended side effect of treating different
consecutive errors like a single one, but that's better than randomly
truncating error messages.