feat(problem_matcher): better highlight file errors. Drop multiline msg. support
The GitHub runner problem matcher is just too limited to achieve the
matching we want properly. Looping pattern hacks are much harder to
maintain, and may skip matching warnings and errors that happen to
follow another error immediately in the output. It's a better tradeoff
to just ignore the second and next lines, but show all the errors and
warnings. Multiline errors and warnings are rare anyway.

While at it, let's highlight file-specific errors in a better way.