refactor(util): simplify getSubmodulePaths function
Instead of parsing the .gitmodules file and reinventing the wheel in the
action code, it's better to ask git directly for the submodule list.

This refactor introduces a slight behavior change: uninitialized
submodules will be ignored. I don't think this is a problem, as if their
folder exists they won't be proper repositories anyway, so we can treat
them as usual.

Part of #54.