perf: improve workflow run modification detection method
It now guards against modification timestamps not representing actual
modifications, which can happen sometimes even when just cloning a
repository, by falling back to a Git object hash check.

In the future, if it's necessary, the action could get a parameter to
select between the hash-only and timestamp + hash modification detection
methods, but for now I think this code is correct and as performant as
possible for practical scenarios.