refactor: rework file modification time set logic (#55)
* refactor: rework file modification time set logic

The previous logic was mostly inherited from a Perl script and adapted
to work on a Typescript environment. It worked acceptably well, barring
the technical debt it ended up introducing.

Let's rewrite it from scratch and document it better.

Part of #54.

* chore(git_set_file_times): rename variable

* chore(git_set_file_times): tweak comment

After more testing, it turns out that Git is smart enough to query
parent repositories from within submodules when the provided paths
belong to parent submodules. However, other gotchas still apply.

* feat: debug log stack trace for unhandled errors

This may come in handy for troubleshooting them.

* chore(git_set_file_times): silence ESLint warning

* chore: tweak some error and debug messages. Add some website refs

* chore(workflows/test_pack_in_submodule): fix typo

* chore: drop @actions/glob dependency

Due to the generated options file being out of GITHUB_WORKSPACE, it
didn't really hash it as we wanted. Moreover, getting the hash of a file
in Node is easy enough that importing a dependency just for this is
overkill.

* fix(util): correct unterminated string error at getSubmodules

The NUL character was interpreted somewhere as a string terminator,
which caused an unterminated string error. Fix that by using a line
break as a separator instead, which should be pretty rare to find in
file paths anyway.

* tweak(test_pack_in_submodule): checkout local action instead

This makes the workflow testable with tools like e.g. act without the
need to push anything anywhere.

* chore(package.json): add build step for submodule test command

* chore(index): make unhandled error handler more concise

* fix(git_set_file_times): correct repo filtering logic

Promises are truthy, so we didn't do any filtering at all with the
filter() call! Moreover, the filtering check was flawed in the case of a
pack directory being in a submodule subfolder.

* chore(dist): build

* test: add test workflow for the "pack repository is a submodule" case

* chore(git_set_file_times): gate extra verbose logging behind env var

This allows interested parties to still get these logs, while not
slowing down things in the most common cases.

* chore(git_set_file_times): correct inverted logic

* perf: improve workflow run modification detection method

It now guards against modification timestamps not representing actual
modifications, which can happen sometimes even when just cloning a
repository, by falling back to a Git object hash check.

In the future, if it's necessary, the action could get a parameter to
select between the hash-only and timestamp + hash modification detection
methods, but for now I think this code is correct and as performant as
possible for practical scenarios.

* chore(dist): build

* fix(git_set_file_times): handle git hash failure for directories

* chore: remove debugging remnant

* test(test_pack_in_submodule): also test dirty file modification behavior

* chore(git_set_file_times): add sanity check for complete Git log parsing

If we stop matching Git log records while we have pending files to set
modification times of, it's a problem that should be dealt with.

* test(test_pack_in_submodule): test worktree file deletion and renaming

* fix(git_set_file_times): better handle directories

Directories could sneak through to the indexed files set otherwise,
which negatively affects performance (potentially more Git log entries
will need to be parsed) and memory consumption. Moreover, by promptly
ignoring them before even checking for modifications, we avoid the need
to handle `git hash-object` erroring out.

* chore(git_set_file_times): extract verbose logging string to constant

* fix(git_set_file_times): correct Git log parsing failure in an edge case