fix: always extract the PackSquash AppImage before executing it
Ubuntu 22 dropped the libfuse2 package, which provides the FUSE 2
implementation the PackSquash v0.3.1 AppImage runtime requires to mount
its filesystem and run. Currently, GitHub is doing a gradual rollout of
Ubuntu 22 for ubuntu-latest runners that will complete on December, so
some repositories already can't use this action, and no repository whose
workflows run on ubuntu-latest will be able to run the action in some
weeks.

To address this pressing issue, let's instruct the action to always
extract the AppImage, as it was done on container environments already.
This does not require FUSE, but it may entail a small performance hit
due to the extraction.