feat: allow overriding `output_file_path`
This allows workflows to instruct the action to save the generated ZIP
to wherever they want to, which is useful to avoid the verbosity and
performance hit of downloading the artifact just after the step
completes.

The default behavior when `output_file_path` is not specified is to
output the generated ZIP to an internal temporary directory.

These changes do not affect artifact generation: for caching and ease of
debugging reasons, it is a good idea to generate artifacts anyway. If
these are too annoying, users can explicitly remove them with other
actions.

Addresses #62.