refactor(ci): drop dependency on PackSquash-action-test repository
This dependency is unnecessary, because that repository was focused on
testing the action, and to test the action we already have some
automated workflows at this repository.

For testing in other environments, the new reusable workflow step
commands can be directly copied and pasted to an Unix-like shell, which
is almost as easy as running `git clone`, but with the added advantage
that at least someone (me) will look at and update this commands
periodically.

While at it, let's update the referenced pack versions to test more
up-to-date packs, which are likely to be more representative of end
PackSquash usage.