feat: support arbitrarily long `PACKSQUASH_SYSTEM_ID`s
https://github.com/ComunidadAylas/PackSquash/commit/bf7188d6ab7a9fe993f10ea9b2fc3dadc3d32486
made PackSquash support arbitrarily long system ID hex strings. In
addition to being somewhat safer and more scalable, such a simple format
also simplifies the action a bit by doing away with the need for an
external dependency for generating UUIDs. However, this new format is
not backwards compatible, so care should be taken to avoid using it with
older PackSquash versions. To achieve that outcome, let's leverage the
version manifest system we already have to introduce such a breaking
behavior bump while handling older versions of both PackSquash and the
action as gracefully as possible.